+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Added tests to check context of called actions
+- Changed message formatter to use its own help method
 
 # 0.1.0